requirements.txt
Django==4.0.2
psycopg2-binary==2.9.3
python-dotenv==0.21.0
django-serverside-datatable==2.0.0
Serverside-tables-django-framework==0.0.3
django-widget-tweaks==1.4.12
django-auditlog==2.2.1
djangorestframework==3.14
openpyxl==3.0.10
requests==2.28.1
pyodbc==4.0.35
django-mssql-backend==2.8.1
Pillow==9.4.0
pyzbar==0.1.9
opencv-python-headless==4.7.0.72
